curl_ntlm_msgs.c: Another attempt to fix compilation warning
curl_ntlm_msgs.c:170: warning: conversion to 'short unsigned int' from
                      'int' may alter its value

+++ lib/curl_ntlm_msgs.c
@@ -166,8 +166,8 @@ static unsigned int readint_le(unsigned char *buf)
  */
 static unsigned short readshort_le(unsigned char *buf)
 {
-  return ((unsigned short)((unsigned short)buf[0]) |
-          (unsigned short)((unsigned short)buf[1] << 8));
+  return (unsigned short)((unsigned short)((unsigned short)buf[0]) |
+                          (unsigned short)((unsigned short)buf[1] << 8));
 }
 
 /*
